+++ src/main/java/co/simplon/usersdao/RoleDao.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import co.simplon.users.Role;
 
-public interface RoleDao extends CrudRepository<Role, Long> {
+public interface RoleDao extends CrudRepository<Role, Integer> {
 	
 	public Role findByName(String name);
